Add Author Split, Add Error Handling
Author Split was added in its own folder and now works with the rest of the program (when ran with a command line shortcut). For some reason it does not work when the folder is with the rest of the control panel code.

The error log that was apart of the open_all window was removed.

All programs in the control panel had error handling added to them. Error message boxes and warning message boxes will now pop up when issues arise.